# api-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Content and Project Management API",
    "description": "API for managing content, projects, categories, subcategories, areas and statuses, with history and access control.",
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "role": { "type": "string", "enum": ["admin", "creator"] },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "Project": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "title": { "type": "string", "maxLength": 200 },
          "description": { "type": "string", "description": "HTML" },
          "target": { "type": "string", "description": "HTML" },
          "conclusion": { "type": "string", "description": "HTML", "nullable": true },
          "priority": { "type": "integer", "minimum": 0, "maximum": 100, "default": 0 },
          "statusId": { "type": "integer" },
          "areaId": { "type": "integer" },
          "authorId": { "type": "integer" },
          "lastUpdatedById": { "type": "integer", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "author": { "$ref": "#/components/schemas/User" },
          "lastUpdatedBy": { "$ref": "#/components/schemas/User" },
          "status": { "$ref": "#/components/schemas/ProjectStatus" },
          "area": { "$ref": "#/components/schemas/ProjectArea" }
        }
      },
      "CreateProjectDto": {
        "type": "object",
        "required": ["title", "description", "target", "statusId", "areaId"],
        "properties": {
          "title": { "type": "string", "minLength": 1, "maxLength": 200 },
          "description": { "type": "string", "description": "HTML" },
          "target": { "type": "string", "description": "HTML" },
          "conclusion": { "type": "string", "description": "HTML", "nullable": true },
          "statusId": { "type": "integer" },
          "areaId": { "type": "integer" },
          "priority": { "type": "integer", "minimum": 0, "maximum": 100 }
        }
      },
      "UpdateProjectDto": {
        "type": "object",
        "properties": {
          "title": { "type": "string", "minLength": 1, "maxLength": 200 },
          "description": { "type": "string", "description": "HTML" },
          "target": { "type": "string", "description": "HTML" },
          "conclusion": { "type": "string", "description": "HTML", "nullable": true },
          "statusId": { "type": "integer" },
          "areaId": { "type": "integer" },
          "priority": { "type": "integer", "minimum": 0, "maximum": 100 }
        }
      },
      "ProjectStatus": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string", "maxLength": 100 },
          "description": { "type": "string" },
          "isActive": { "type": "boolean", "default": true },
          "priority": { "type": "integer", "minimum": 0, "maximum": 100, "default": 0 },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "CreateProjectStatusDto": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 100 },
          "description": { "type": "string" },
          "isActive": { "type": "boolean" },
          "priority": { "type": "integer", "minimum": 0, "maximum": 100 }
        }
      },
      "UpdateProjectStatusDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 100 },
          "description": { "type": "string" },
          "isActive": { "type": "boolean" },
          "priority": { "type": "integer", "minimum": 0, "maximum": 100 }
        }
      },
      "ProjectArea": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string", "maxLength": 100 },
          "description": { "type": "string" },
          "isActive": { "type": "boolean", "default": true },
          "priority": { "type": "integer", "minimum": 0, "maximum": 100, "default": 0 },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "CreateProjectAreaDto": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 100 },
          "description": { "type": "string" },
          "isActive": { "type": "boolean" },
          "priority": { "type": "integer", "minimum": 0, "maximum": 100 }
        }
      },
      "UpdateProjectAreaDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 100 },
          "description": { "type": "string" },
          "isActive": { "type": "boolean" },
          "priority": { "type": "integer", "minimum": 0, "maximum": 100 }
        }
      },
      "ProjectHistory": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "action": { "type": "string", "enum": ["created", "updated", "status_changed", "area_changed", "conclusion_added", "deleted"] },
          "changes": { "type": "string" },
          "notes": { "type": "string", "nullable": true },
          "previousData": { "type": "string", "nullable": true },
          "newData": { "type": "string", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "user": { "$ref": "#/components/schemas/User" },
          "project": { "$ref": "#/components/schemas/Project" }
        }
      }
    }
  },
  "paths": {
...
    "/projects": {
      "get": {
        "tags": ["Proyectos"],
        "summary": "Listar proyectos",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "orderBy", "in": "query", "schema": { "type": "string", "enum": ["priority", "createdAt", "updatedAt", "title"], "default": "priority" } },
          { "name": "orderDirection", "in": "query", "schema": { "type": "string", "enum": ["ASC", "DESC"], "default": "DESC" } },
          { "name": "statusId", "in": "query", "schema": { "type": "integer" } },
          { "name": "areaId", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Lista de proyectos",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Project" } } } }
          }
        }
      },
      "post": {
        "tags": ["Proyectos"],
        "summary": "Crear proyecto",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateProjectDto" } } }
        },
        "responses": {
          "201": {
            "description": "Proyecto creado",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Project" } } }
          }
        }
      }
    },
    "/projects/{id}": {
      "get": {
        "tags": ["Proyectos"],
        "summary": "Obtener proyecto por ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } } ],
        "responses": {
          "200": { "description": "Proyecto encontrado", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Project" } } } },
          "404": { "description": "No encontrado" }
        }
      },
      "patch": {
        "tags": ["Proyectos"],
        "summary": "Actualizar proyecto",
        "security": [{ "bearerAuth": [] }],
        "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } } ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateProjectDto" } } }
        },
        "responses": {
          "200": { "description": "Proyecto actualizado", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Project" } } } }
        }
      },
      "delete": {
        "tags": ["Proyectos"],
        "summary": "Eliminar proyecto",
        "security": [{ "bearerAuth": [] }],
        "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } } ],
        "responses": {
          "200": { "description": "Proyecto eliminado" }
        }
      }
    },
    "/projects/{id}/history": {
      "get": {
        "tags": ["Proyectos"],
        "summary": "Historial de un proyecto",
        "security": [{ "bearerAuth": [] }],
        "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } } ],
        "responses": {
          "200": { "description": "Historial del proyecto", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/ProjectHistory" } } } } }
        }
      }
    },
    "/projects/history/user/{userId}": {
      "get": {
        "tags": ["Proyectos"],
        "summary": "Historial de proyectos de un usuario (solo admin)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [ { "name": "userId", "in": "path", "required": true, "schema": { "type": "integer" } } ],
        "responses": {
          "200": { "description": "Historial del usuario", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/ProjectHistory" } } } } }
        }
      }
    },
    "/projects/history/my": {
      "get": {
        "tags": ["Proyectos"],
        "summary": "Historial de mis proyectos",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Mi historial de proyectos", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/ProjectHistory" } } } } }
        }
      }
    },
    "/project-statuses": {
      "get": {
        "tags": ["Estados de Proyecto"],
        "summary": "Listar todos los estados (admin)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "active", "in": "query", "schema": { "type": "string", "enum": ["true", "false"] } },
          { "name": "orderBy", "in": "query", "schema": { "type": "string", "enum": ["priority", "name", "createdAt", "updatedAt"], "default": "priority" } },
          { "name": "orderDirection", "in": "query", "schema": { "type": "string", "enum": ["ASC", "DESC"], "default": "DESC" } }
        ],
        "responses": {
          "200": { "description": "Lista de estados", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/ProjectStatus" } } } } }
        }
      },
      "post": {
        "tags": ["Estados de Proyecto"],
        "summary": "Crear estado (admin)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateProjectStatusDto" } } }
        },
        "responses": {
          "201": { "description": "Estado creado", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProjectStatus" } } } }
        }
      }
    },
    "/project-statuses/public": {
      "get": {
        "tags": ["Estados de Proyecto"],
        "summary": "Listar estados activos (público)",
        "parameters": [
          { "name": "orderBy", "in": "query", "schema": { "type": "string", "enum": ["priority", "name", "createdAt", "updatedAt"], "default": "priority" } },
          { "name": "orderDirection", "in": "query", "schema": { "type": "string", "enum": ["ASC", "DESC"], "default": "DESC" } }
        ],
        "responses": {
          "200": { "description": "Lista de estados activos", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/ProjectStatus" } } } } }
        }
      }
    },
    "/project-statuses/{id}": {
      "get": {
        "tags": ["Estados de Proyecto"],
        "summary": "Obtener estado por ID (admin)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } } ],
        "responses": {
          "200": { "description": "Estado encontrado", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProjectStatus" } } } }
        }
      },
      "patch": {
        "tags": ["Estados de Proyecto"],
        "summary": "Actualizar estado (admin)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } } ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateProjectStatusDto" } } }
        },
        "responses": {
          "200": { "description": "Estado actualizado", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProjectStatus" } } } }
        }
      },
      "delete": {
        "tags": ["Estados de Proyecto"],
        "summary": "Eliminar estado (admin)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } } ],
        "responses": {
          "200": { "description": "Estado eliminado" }
        }
      }
    },
    "/project-statuses/{id}/toggle": {
      "patch": {
        "tags": ["Estados de Proyecto"],
        "summary": "Activar/desactivar estado (admin)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } } ],
        "responses": {
          "200": { "description": "Estado actualizado", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProjectStatus" } } } }
        }
      }
    },
    "/project-areas": {
      "get": {
        "tags": ["Áreas de Proyecto"],
        "summary": "Listar todas las áreas (admin)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "active", "in": "query", "schema": { "type": "string", "enum": ["true", "false"] } },
          { "name": "orderBy", "in": "query", "schema": { "type": "string", "enum": ["priority", "name", "createdAt", "updatedAt"], "default": "priority" } },
          { "name": "orderDirection", "in": "query", "schema": { "type": "string", "enum": ["ASC", "DESC"], "default": "DESC" } }
        ],
        "responses": {
          "200": { "description": "Lista de áreas", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/ProjectArea" } } } } }
        }
      },
      "post": {
        "tags": ["Áreas de Proyecto"],
        "summary": "Crear área (admin)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateProjectAreaDto" } } }
        },
        "responses": {
          "201": { "description": "Área creada", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProjectArea" } } } }
        }
      }
    },
    "/project-areas/public": {
      "get": {
        "tags": ["Áreas de Proyecto"],
        "summary": "Listar áreas activas (público)",
        "parameters": [
          { "name": "orderBy", "in": "query", "schema": { "type": "string", "enum": ["priority", "name", "createdAt", "updatedAt"], "default": "priority" } },
          { "name": "orderDirection", "in": "query", "schema": { "type": "string", "enum": ["ASC", "DESC"], "default": "DESC" } }
        ],
        "responses": {
          "200": { "description": "Lista de áreas activas", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/ProjectArea" } } } } }
        }
      }
    },
    "/project-areas/{id}": {
      "get": {
        "tags": ["Áreas de Proyecto"],
        "summary": "Obtener área por ID (admin)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } } ],
        "responses": {
          "200": { "description": "Área encontrada", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProjectArea" } } } }
        }
      },
      "patch": {
        "tags": ["Áreas de Proyecto"],
        "summary": "Actualizar área (admin)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } } ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateProjectAreaDto" } } }
        },
        "responses": {
          "200": { "description": "Área actualizada", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProjectArea" } } } }
        }
      },
      "delete": {
        "tags": ["Áreas de Proyecto"],
        "summary": "Eliminar área (admin)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } } ],
        "responses": {
          "200": { "description": "Área eliminada" }
        }
      }
    },
    "/project-areas/{id}/toggle": {
      "patch": {
        "tags": ["Áreas de Proyecto"],
        "summary": "Activar/desactivar área (admin)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } } ],
        "responses": {
          "200": { "description": "Área actualizada", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProjectArea" } } } }
        }
      }
    }
  }
} 